+++ websubsub/management/commands/websub_purge_unresolvable.py
@@ -13,7 +13,14 @@ log = logging.getLogger('websubsub')
 
 class Command(BaseCommand):
     help = 'Delete all subscriptions with unresolvable urlname from database.'
-
+    
+    def add_arguments(self, parser):
+        parser.add_argument(
+            '-y', '--yes',
+            action='store_true',
+            help='answer yes to all',
+        )
+        
     def handle(self, *args, **kwargs):
         unresolvable_pks = set()
         for ssn in Subscription.objects.all():
@@ -29,13 +36,14 @@ class Command(BaseCommand):
             print('No unresolvable subscriptions found')
             return
         
-        text = f'Are you sure you want to delete {len(unresolvable_pks)} subscriptions? (y/N): '
-        while True:
-            answer = input(text)
-            if not answer or answer in ('n','N'):
-                return
-            if answer in ('y', 'Y'):
-                break
+        if not kwargs['yes']:
+            text = f'Are you sure you want to delete {len(unresolvable_pks)} subscriptions? (y/N): '
+            while True:
+                answer = input(text)
+                if not answer or answer in ('n','N'):
+                    return
+                if answer in ('y', 'Y'):
+                    break
         
         Subscription.objects.filter(pk__in=unresolvable_pks).delete()
         print(f'{len(unresolvable_pks)} subscriptions was successfully removed from database')
